Removes some signed/unsigned comparison warnings

+++ libi3ds/src/gige_camera_sensor.cpp
@@ -491,8 +491,8 @@ i3ds::GigECamera::send_sample(const byte* image, int width, int height)
   // BOOST_LOG_TRIVIAL(info) << "size check x: " << width << " " << frame.descriptor.region.size_x;
   // BOOST_LOG_TRIVIAL(info) << "size check y: " << height <<" " << frame.descriptor.region.size_y;
   if(
-      ( width  != frame.descriptor.region.size_x ) ||
-      ( height != frame.descriptor.region.size_y )
+      ( (unsigned int)width  != frame.descriptor.region.size_x ) ||
+      ( (unsigned int)height != frame.descriptor.region.size_y )
       )
     {
       BOOST_LOG_TRIVIAL(error) << "Error in image format going to be sent";
